Oops, forgot to commit a bunch of work from a week or two ago
Add preconditioner to CG declip filter


git-svn-id: https://svn.xiph.org/trunk/postfish@5759 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ reconstruct.h
@@ -21,5 +21,5 @@
  * 
  */
 
-extern void reconstruct(drft_lookup *fft,double *x, double *w, int *flag, 
+extern void reconstruct(drft_lookup *fft,double *x, double *w, double *flag, 
 			double e,int max,int n);
